requirements.txt
fastapi
uvicorn
requests
python-dotenv
PyMuPDF
SQLAlchemy
celery
redis
pytest
python-multipart
